Move ExAuth, FKOAuth1 & FKOAuthDataStore to own namespace `Friendica\Security`
