+++ src/Security/FKOAuth1.php
@@ -24,6 +24,7 @@ namespace Friendica\Network;
 use Friendica\Core\Logger;
 use Friendica\Database\DBA;
 use Friendica\DI;
+use Friendica\Security\FKOAuthDataStore;
 use OAuthServer;
 use OAuthSignatureMethod_HMAC_SHA1;
 use OAuthSignatureMethod_PLAINTEXT;
